Replaced homepage images with .webp
