--- packages/Webkul/Product/src/CacheFilters/Large.php
@@ -1,28 +1,0 @@
-<?php
-
-namespace Webkul\Product\CacheFilters;
-
-use Intervention\Image\Image;
-use Intervention\Image\Filters\FilterInterface;
-
-class Large implements FilterInterface
-{
-    /**
-     * Apply filter.
-     *
-     * @param  \Intervention\Image\Image  $image
-     * @return \Intervention\Image\Image
-     */
-    public function applyFilter(Image $image)
-    {
-        $width = core()->getConfigData('catalog.products.cache_large_image.width') != ''
-            ? core()->getConfigData('catalog.products.cache_large_image.width')
-            : 560;
-
-        $height = core()->getConfigData('catalog.products.cache_large_image.height') != ''
-            ? core()->getConfigData('catalog.products.cache_large_image.height')
-            : 610;
-
-        return $image->fit($width, $height);
-    }
-}
